fix unintuitive default input mapping to retropad
@shakkalakka has prepared extensive notes on the issue of the unintuitive/difficult to use default input mapping of this core: https://forums.libretro.com/t/default-input-mapping-for-mame2014-and-newer-is-wrong/11281

I'm submitting this PR on behalf of shakkalakka because they are concerned about their language barrier.

Many of the popular games already have coded independent defaults farther down in the source. In other words, this default does not change how every game is handled.

One way to evaluate the change would be to try to come up with any prominent examples where this would not be more intuitive.